feat: Add Navbar components and AI tools pages
- Implemented NavbarBottom and NavbarTop components for improved navigation.
- Created AIDashboard, AIGenerate, AIImage, BackgroundRemover, BlogTitle, ObjectRemover, ResumeReviewer, and WriteArticles pages for AI tools.
- Developed AITools and HomePage components to showcase AI tools and testimonials.
- Enhanced UI with responsive design and improved user experience across various components.

+++ client/src/pages/ShoppingCartPage.jsx
@@ -1,19 +1,24 @@
-import { NavLink } from "react-router-dom";
-import { ShoppingCart } from "lucide-react";
-import { NotFound } from "../utils";
-import { CartListing, HomeCertificate } from "../components";
-import { useState } from "react";
+import { NavLink, useNavigate } from "react-router-dom";
+import { ShoppingCart, Trash2Icon } from "lucide-react";
+import { Loading, NotFound } from "../utils";
+import { HomeCertificate } from "../components";
+import { useEffect, useState } from "react";
 import { toast } from "react-toastify";
 import { axios, errorHandler } from "../helper";
 import { useDispatch, useSelector } from "react-redux";
-import { removeItem, updateItemQuantity } from "../redux/cartSlice";
+import { fetchCarts, removeItem, updateItemQuantity } from "../redux/cartSlice";
 
 const ShoppingCartPage = () => {
-  const { items } = useSelector((state) => state.cart);
+  const { items, loading } = useSelector((state) => state.cart);
   const { items: addresses } = useSelector((state) => state.address);
   const dispatch = useDispatch();
   const [selectAddress, setSelectAddress] = useState(null);
   const { user } = useSelector((s) => s.auth);
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    dispatch(fetchCarts());
+  }, [dispatch]);
 
   const handleCheckout = async () => {
     try {
@@ -33,9 +38,11 @@ const ShoppingCartPage = () => {
   const totals =
     items && items?.reduce((p, c) => c?.productId?.price * c?.quantity + p, 0);
 
+  if (loading) return <Loading />;
+
   return (
     <section className="min-h-screen sm:p-4 p-3 max-w-6xl mx-auto">
-      {items.length === 0 && (
+      {items?.length === 0 && (
         <NotFound
           canvas={
             <ShoppingCart className="w-20 h-20 text-gray-400 mb-4 mx-auto" />
@@ -51,24 +58,26 @@ const ShoppingCartPage = () => {
       <div className="flex max-sm:flex-col gap-5">
         <div className="md:flex-1 w-full">
           {items &&
-            items.map((item) => (
-              <CartListing
-                key={item._id}
-                {...item}
-                onQtyChange={(quantity) =>
-                  dispatch(updateItemQuantity({ _id: item._id, quantity }))
-                }
-                onRemove={() => dispatch(removeItem(item._id))}
-              />
-            ))}
+            items?.map((item) => {
+              if (!item?.productId) return null;
+              return (
+                <CartListing
+                  key={item._id}
+                  {...item}
+                  onQtyChange={(quantity) =>
+                    dispatch(updateItemQuantity({ _id: item._id, quantity }))
+                  }
+                  onRemove={() => dispatch(removeItem(item._id))}
+                />
+              );
+            })}
 
           <p className="font-semibold text-2xl py-4">Shipping Address</p>
 
           <div
-            className={`capitalize border border-gray-300 cursor-pointer !px-5 mb-4 card !bg-transparent`}>
-            <NavLink to={"/shipping"} className="font-semibold">
-              Add New Address
-            </NavLink>
+            onClick={() => navigate("/shipping-address")}
+            className="font-semibold capitalize border border-gray-300 cursor-pointer !px-5 mb-4 card !bg-transparent">
+            Add New Address
           </div>
 
           {addresses?.map((i) => (
@@ -86,7 +95,7 @@ const ShoppingCartPage = () => {
             </div>
           ))}
 
-          {selectAddress && (
+          {selectAddress && !!items?.length && (
             <div className="flex gap-6 font-semibold mt-10 ">
               <NavLink
                 to={"/product"}
@@ -102,25 +111,27 @@ const ShoppingCartPage = () => {
           )}
         </div>
 
-        <div className="p-4 md:w-1/3 space-y-3 w-full sticky top-20 h-fit">
-          <h3>This Order shipping Fee!</h3>
-          <div className="flex justify-between font-semibold text-xl">
-            <span>({items?.length}) Item</span>
-            <span>${totals | 1}</span>
+        {!!items?.length && (
+          <div className="p-4 md:w-1/3 space-y-3 w-full sticky top-20 h-fit">
+            <h3>This Order shipping Fee!</h3>
+            <div className="flex justify-between font-semibold text-xl">
+              <span>({items?.length}) Item</span>
+              <span>${totals | 1}</span>
+            </div>
+            <div className="flex justify-between">
+              <span>Shipping:</span>
+              <span>FREE</span>
+            </div>
+            <div className="flex justify-between">
+              <span>Estimate Tax:</span>
+              <span>$5</span>
+            </div>
+            <div className="flex justify-between font-semibold text-3xl">
+              <span>Total:</span>
+              <span>${((totals + 5) | 0).toFixed(2)}</span>
+            </div>
           </div>
-          <div className="flex justify-between">
-            <span>Shipping:</span>
-            <span>FREE</span>
-          </div>
-          <div className="flex justify-between">
-            <span>Estimate Tax:</span>
-            <span>$5</span>
-          </div>
-          <div className="flex justify-between font-semibold text-3xl">
-            <span>Total:</span>
-            <span>${((totals + 5) | 0).toFixed(2)}</span>
-          </div>
-        </div>
+        )}
       </div>
 
       <HomeCertificate />
@@ -129,3 +140,83 @@ const ShoppingCartPage = () => {
 };
 
 export default ShoppingCartPage;
+
+const CartListing = ({ productId, quantity, onRemove, onQtyChange }) => {
+  const { _id, thumbnail, title, price, category, brand } = productId;
+  const [qty, setQty] = useState(quantity);
+
+  const handleDelete = async () => {
+    try {
+      if (!_id) return;
+      const res = await axios.delete(`/cart/${_id}`);
+      if (res.data) {
+        onRemove();
+      }
+    } catch (error) {
+      errorHandler(error);
+    }
+  };
+  const handleUpdateQty = async () => {
+    try {
+      if (!_id || !qty) return;
+      const res = await axios.put(`/cart`, {
+        productId: _id,
+        quantity: qty,
+      });
+      if (res.data) {
+        onQtyChange(qty);
+      }
+    } catch (error) {
+      errorHandler(error);
+    }
+  };
+
+  return (
+    <div className="flex max-sm:flex-col gap-5 items-start border-b text-slate-700 border-gray-300 py-4">
+      <NavLink to={`/product/${_id}`} className="bg-gray-300 max-sm:w-full">
+        <img
+          src={thumbnail || "https://placehold.co/120x120"}
+          alt="Product"
+          className="w-[200px] mx-auto object-cover rounded transition-opacity duration-300 opacity-100"
+        />
+      </NavLink>
+      <div className="sm:flex-1 max-sm:px-4 w-full space-y-2 capitalize">
+        <p className="font-medium text-lg">{title || "Smartphone X Pro"}</p>
+        <p>Category : {category || "other"}</p>
+        <p>Brand: {brand || "other"}</p>
+        <div className="flex items-center my-1">
+          <span>Unit Price:</span>
+          <span className="ml-2 font-semibold">
+            ${price || 79.99} x {qty}
+          </span>
+        </div>
+        <p>
+          Totals: <span className="font-bold">${price * qty}</span>
+        </p>
+        <div className="flex gap-5 items-center mt-3">
+          <div className="py-1 gap-6 flex items-center justify-center px-6 border border-slate-300 rounded-full w-fit font-medium">
+            <button
+              className="text-center text-xl"
+              onClick={() => {
+                if (qty !== 1) setQty((prev) => prev - 1);
+              }}>
+              -
+            </button>
+            <button className="text-center">{qty}</button>
+            <button
+              className="text-center text-xl"
+              onClick={() => {
+                if (qty < 5) setQty((prev) => prev + 1);
+              }}>
+              +
+            </button>
+            {quantity !== qty && <button onClick={handleUpdateQty}>Set</button>}
+          </div>
+          <button onClick={handleDelete} className="text-red-600 svg-btn !p-2">
+            <Trash2Icon />
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+};
